renaming lib dir so doesn't clash in wrapper
